move export-responses as single component to its right location
